Update calls to sdo_geometry to fix ORA-13343
Prior calls to sdo_geometry specified a 2D polygon geometry (2003)
specified as an outer ring (1003). As only two values are being passed
into the ordinate array, this creates an invalid polygon, with fewer
than 4 points. In some cases, this can trigger an ORA-13343.

Update the call to sdo_geometry to specify a 2D point (2001) and a point
type elem_info_array, as we are passing in 2 values (long, lat) that
specify a point, not a polygon.