[testcontainers] modified to allow container bypass.
It's been frustrating to diagnose why test harnesses and such fail. A
mechanism to let the container object go directly to an existing
database was put in place to make things like initial test
infrastructure creation easier. It is not a replacement for actually
asserting though required calls the state of the database that the tests
need.