Updated docker install for test environments.
Allows for repeated calls to build that are idempotent by check the
installed version vs what's in the schema.

NOTE: it makes no attempt to upgrade if the versions don't match just
runs the install. This will work better when we're using flyway as it
will directly check the version run the required updates.