Fix regression with adding VPD on av_sec_users.
The av_sec_users view now has a VPD policy on it so that PD/DBA Users
can see and use the view. Unfortunately the get_assigned_privs function
also used that view. So when the login trigger fired it call
get_assigned_privs and queried the view, which in turned triggered
get_assigned_privs.

Changing the order of the check in the VDP handler so the CWMS_20
context is checked first prevents the needless check of the view in that
case.